Complete Update products feature

+++ shop/products/forms.py
@@ -1,10 +1,10 @@
-from flask_wtf.file import FileAllowed, FileField, FileRequired
-from wtforms import Form, IntegerField, StringField, TextAreaField, validators
+from flask_wtf.file import FileAllowed, FileField
+from wtforms import Form, IntegerField, DecimalField, StringField, TextAreaField, validators
 
 class ProductForm(Form):
     name = StringField("Name", [validators.DataRequired()])
     origins = StringField("Country of Origins", [validators.DataRequired()])
-    price = IntegerField("Price", [validators.DataRequired()])
+    price = DecimalField("Price", [validators.DataRequired()])
     discount = IntegerField("Discount", default=0)
     stock = IntegerField("Stock", [validators.DataRequired()])
     description = TextAreaField("Description", [validators.DataRequired()])
